Chore: pre-release deployment
Update the github action a manually triggered build job using workflow_dispatch.
If the user selects a branch to deploy from all build artifacts will be commited to the same branch and the new (release) commit will be tagged with the user input. If on the other hand the user selects a git tag to deploy from we will create a new branch and a pull request to be reviewed by a team member.